Correcting the CodedInputStream-reading routines.
The 'length' value indicates the number of bytes for this section of the input
stream, but because values may be encoded using fewer bytes than sizeof(type),
we can't use 'length' to compute a count.

This invalidates the previous memory allocation optimization we were
attempting, so we now unfortunately need to occasionally realloc() while
adding values from the input stream.