Removing the CodedInputStream routines from PBArray.
Previously, we were only supporting a subset of the CodedInputStream reading
routines.  In order to support types like Fixed32, we'd need to further expand
our set of array value types simply to recognize these additional stream
reading functions.

Instead, we revert to the earlier approach of hanging all stream operations
externally.  There may be a small performance and code-size penalty in doing
so, but those can be revisited some time in the future.