Support the PBMessage_Builder protocol.
PBUnknownFieldSet_Builder needs to conform to the PBMessage_Builder
protocol in order for it to be used in all of the places expecting
id<PBMessage_Builder> values.

The protocol itself is pretty heavy, and some of the methods could be
made @optional, but for now we just raise exceptions for unsupported
calls (which is consistent with the rest of the runtime library).

This fixes the Apple LLVM compiler 3.0 warnings reported in #16.