avoid "local ... hides instance variable" warnings in generated code
GCC gets upset if you hide an instance variable with a local variable
and issues a warning that apparently can't be turned off.  This commit
renames the "size" local variable in the serializedSize generated
method so it doesn't hide the instance variable.  It also renames the
"value" parameter to the setHasFoo: methods to avoid hiding a field
named "value" if it happens to exist.